Add toggleable support to RadioListTile, fix RadioListTile and SwitchListTile tests.
